refactor: InputView
