+++ README.md
@@ -7,3 +7,12 @@
 
 ## 온라인 코드 리뷰 과정
 * [텍스트와 이미지로 살펴보는 온라인 코드 리뷰 과정](https://github.com/next-step/nextstep-docs/tree/master/codereview)
+
+## step3
+### 기능 구현할 목록 정리
+1. 사용자는 몇 대의 자동차로 몇 번의 이동을 할 것인지를 입력할 수 있어야 한다.(InputView에서 구현, UI이므로 단위 테스트X)
+2. 입력한 자동차의 대수에 따라 자동차 객체 생성 하기
+3. 생성한 객체 배열에 담기(자동차의 대수 = 배열의 크기)
+4. 전진하는 조건(상수)를 받아 4보다 같거나 클 경우 전진(+1), 작을경우(+0) 하는 로직 구현
+5. 전진하는 조건(random)값으로 받기(random 받는 부분이므로 단위 테스트X)
+6. 자동차의 상태를 화면에 출력한다. 어느 시점에 출력할 것인지에 대한 제약은 없다.(ResultView에서 구현, UI이므로 단위 테스트X)
