+++ src/test/java/study/calculator/SetTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.calculator;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
