+++ src/test/java/study/calculator/StringAddCalculatorTest.java
@@ -1,7 +1,8 @@
-package study;
+package study.calculator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import study.calculator.Calculator;
 
 import static org.assertj.core.api.Assertions.*;
