+++ src/test/java/study/race/CarRacingTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.race;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
